refactor 각 스캔 역할 분리
