+++ src/main/java/core/di/factory/ClassPathBeanScanner.java
@@ -3,13 +3,13 @@ package core.di.factory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ClassPathBeanScanner1 {
+public class ClassPathBeanScanner {
 
-    private static final Logger logger = LoggerFactory.getLogger(ClassPathBeanScanner1.class);
+    private static final Logger logger = LoggerFactory.getLogger(ClassPathBeanScanner.class);
 
     private final BeanFactory beanFactory;
 
-    public ClassPathBeanScanner1(BeanFactory beanFactory) {
+    public ClassPathBeanScanner(BeanFactory beanFactory) {
         this.beanFactory = beanFactory;
     }
